requirements.txt
Django==3.1.1
django-filter==2.4.0
djangorestframework==3.12.1
requests==2.24.0
